Fix some line endings (I'm on linux now), add Quandrix sort
Quandrix sort is a proof of concept for a Magic: The Gathering combo.
The combo in question is Kalamax + Quandrix Apprentice + A copy spell to allow a player to sort their library.
The sort works by taking chunks of 3 elements and bubble sorting them.
The array must not have a length divisible by 3.